added footer and changed picture
I have integrated a footer into the HTML and styled it using CSS, including the addition of an image. Additionally, I have adjusted the transparency of the Bits College image and applied the same effect to the header. To ensure a seamless layout, I have set the page corners to be fully attached by utilizing margin: 0. These modifications can be found at the top of the CSS file for partners.html.